Removed (wrong) workaround for alignment of pathval storage and ensured the correct behaviour by setting alignemnt of the type.

+++ src/xnode--0.7.1.sql
@@ -97,7 +97,7 @@ CREATE TYPE pathval (
 	internallength = variable,
 	input = pathval_in,
 	output =pathval_out,
-	alignment = int,
+	alignment = double,
 	storage = extended
 );
